Move ComputeManager to use spawn_on
So far the ComputeManager._sync_power_state functionality used an
explicit GreenPool. We change this pool to a GreenThreadPoolExecutor
and use then spawn_on to start the thread. This way we can
* can replace the direct eventlet dependency with an indirect one behind
  an interface where the impl can easily switched out to
  ThreadPoolExecutor in the future
* we can get rid of the explicit pass_context call and centralize that
  instead to spawn_*

The ComputeManager live_migration_executor is already a
GreenThreaPoolExecutor so here we only switching to spawn_on to get rid
of the explicit pass_context call.

Change-Id: I51c5339d3f54f5c66856b6d4e06cd91ac04977cb