[vncproxy]Handle ssl.wrap_socket removal in py312
The ssl.wrap_socket was deprecated in py3.7 and was removed in py3.12 [1].
Instead we have to use ssl.create_default_context[2] to get a context first
and then use the wrap_socket from the context.

This removal was hidden by eventlet monkey patching as that still
introducing the global wrapper. Now that we are trying to remove
eventlet and running the tests without eventlet monkey patching
ssl.wrap_socket simply fails.

So this patch translates the code to use the new way to wrap a socket
for SSL.

[1] https://docs.python.org/3.12/whatsnew/3.12.html#ssl
[2] https://docs.python.org/3/library/ssl.html#socket-creation

Change-Id: I59517df30481fb9bc7ae135c9cabe3b4be39684c
Signed-off-by: Balazs Gibizer <gibi@redhat.com>